refactor : API 호출 hook 위치변경
