+++ src/pages/TransactionsPage.tsx
@@ -1,5 +1,5 @@
 import Loader from '@components/template/Loader';
-import { useTransactions } from '@hooks/useTransaction';
+import { useTransactions } from '@hooks/api/useTransaction';
 import Transactions from '@components/template/Transactions';
 import Error from '@components/template/Error';
 import { useAuthStore } from '@store/authStore';
